+++ plopfile.ts
@@ -4,13 +4,20 @@ import path from 'node:path'
 import type { NodePlopAPI } from 'plop'
 
 type Answers = {
-  layer: 'core' | 'kits' | 'ui'
-  subLayer?: 'hooks' | 'utils'
+  layer: 'core' | 'hooks' | 'kits' | 'ui' | 'utils'
   uiSubLayer?: string
   name: string
 }
 
 export default function (plop: NodePlopAPI) {
+  plop.setHelper('depth', (_, options) => {
+    const answers = options.data.root
+    const fullPath = [answers.layer, answers.uiSubLayer].filter(Boolean).join('/')
+    const parts = fullPath.split('/').length + 2
+
+    return '../'.repeat(parts)
+  })
+
   plop.setHelper('capitalize', (text: string) => text.charAt(0).toUpperCase() + text.slice(1))
 
   plop.setHelper('snakeCase', (text: string) =>
@@ -27,14 +34,7 @@ export default function (plop: NodePlopAPI) {
         type: 'list',
         name: 'layer',
         message: 'Which layer?',
-        choices: ['core', 'kits', 'ui']
-      },
-      {
-        type: 'list',
-        name: 'subLayer',
-        message: 'Choose sublayer (only for core)',
-        when: (answers) => answers.layer === 'core',
-        choices: ['hooks', 'utils']
+        choices: ['core', 'hooks', 'kits', 'ui', 'utils']
       },
       {
         type: 'list',
@@ -60,62 +60,62 @@ export default function (plop: NodePlopAPI) {
     actions: [
       {
         type: 'add',
-        path: 'packages/{{layer}}{{#if subLayer}}/{{subLayer}}{{/if}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/package.json',
+        path: 'packages/{{layer}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/package.json',
         templateFile: 'plop-templates/package.json.hbs'
       },
       {
         type: 'add',
-        path: 'packages/{{layer}}{{#if subLayer}}/{{subLayer}}{{/if}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/tsconfig.json',
+        path: 'packages/{{layer}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/tsconfig.json',
         templateFile: 'plop-templates/tsconfig.json.hbs'
       },
       {
         type: 'add',
-        path: 'packages/{{layer}}{{#if subLayer}}/{{subLayer}}{{/if}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/tsup.config.ts',
+        path: 'packages/{{layer}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/tsup.config.ts',
         templateFile: 'plop-templates/tsup.config.ts.hbs'
       },
       {
         type: 'add',
-        path: 'packages/{{layer}}{{#if subLayer}}/{{subLayer}}{{/if}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/README.md',
+        path: 'packages/{{layer}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/README.md',
         templateFile: 'plop-templates/README.md.hbs'
       },
       {
         type: 'add',
-        path: 'packages/{{layer}}{{#if subLayer}}/{{subLayer}}{{/if}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/src/index.ts',
+        path: 'packages/{{layer}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/src/index.ts',
         templateFile: 'plop-templates/index.ts.hbs',
         skip: (answers: Answers) =>
           answers.layer === 'ui' ? 'UI packages have no empty index' : false
       },
       {
         type: 'add',
-        path: 'packages/{{layer}}{{#if subLayer}}/{{subLayer}}{{/if}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/src/index.ts',
+        path: 'packages/{{layer}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/src/index.ts',
         templateFile: 'plop-templates/ui-index.ts.hbs',
         skip: (answers: Answers) =>
           answers.layer !== 'ui' ? 'Default UI index is only available for UI packages' : false
       },
       {
         type: 'add',
-        path: 'packages/{{layer}}{{#if subLayer}}/{{subLayer}}{{/if}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/src/{{capitalize name}}.tsx',
+        path: 'packages/{{layer}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/src/{{capitalize name}}.tsx',
         templateFile: 'plop-templates/component.tsx.hbs',
         skip: (answers: Answers) =>
           answers.layer !== 'ui' ? 'Components are only available for UI packages' : false
       },
       {
         type: 'add',
-        path: 'packages/{{layer}}{{#if subLayer}}/{{subLayer}}{{/if}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/__stories__/{{capitalize name}}.stories.tsx',
+        path: 'packages/{{layer}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/__stories__/{{capitalize name}}.stories.tsx',
         templateFile: 'plop-templates/component.stories.tsx.hbs',
         skip: (answers: Answers) =>
           answers.layer !== 'ui' ? 'Stories are only available for UI packages' : false
       },
       {
         type: 'add',
-        path: 'packages/{{layer}}{{#if subLayer}}/{{subLayer}}{{/if}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/__tests__/{{capitalize name}}.spec.tsx',
+        path: 'packages/{{layer}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/__tests__/{{capitalize name}}.spec.tsx',
         templateFile: 'plop-templates/component.spec.tsx.hbs',
         skip: (answers: Answers) =>
           answers.layer !== 'ui' ? 'Use ts spec for non-ui packages' : false
       },
       {
         type: 'add',
-        path: 'packages/{{layer}}{{#if subLayer}}/{{subLayer}}{{/if}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/__tests__/{{snakeCase name}}.spec.ts',
+        path: 'packages/{{layer}}{{#if uiSubLayer}}/{{uiSubLayer}}{{/if}}/{{name}}/__tests__/{{snakeCase name}}.spec.ts',
         templateFile: 'plop-templates/component.spec.ts.hbs',
         skip: (answers: Answers) =>
           answers.layer === 'ui' ? 'Use tsx spec for UI packages' : false